fix(#231): extract multiple type definitions from single files into separate files
- MedicationViewModel.swift: split into MedicationViewModel, AddMedicationViewModel, LogMedicationDoseViewModel
- InsightsDashboardView.swift: extract 8 supporting views into InsightsDashboardComponents.swift
- DashboardView.swift: extract 5 supporting views into DashboardComponents.swift
- SymptomInfoViews.swift: move from ViewModels/Symptom/ to Views/Bowel/ and split into BristolStoolInfoView, PainLevelInfoView, UrgencyLevelInfoView, SymptomInfoComponents
- ContentView.swift: extract TabBarAction enum to Navigation/TabBarAction.swift
- AppRouter.swift: extract AppDestination and SheetDestination to separate files
- CustomTabBar.swift: extract TabBarItem to separate file
- SymptomEditView.swift: extract PainLevel extension to Extensions/PainLevel+Conversion.swift
