+++ GutCheck/GutCheck/Navigation/SheetDestination.swift
@@ -1,0 +1,23 @@
+//
+//  SheetDestination.swift
+//  GutCheck
+//
+//  Sheet presentation destinations for the app.
+//
+
+// Sheet presentations
+enum SheetDestination: Identifiable {
+    case profile
+    case mealForm(String? = nil) // nil for new, ID for edit
+    case symptomForm(String? = nil) // nil for new, ID for edit
+    case logEntry // Entry point for choosing what to log
+    
+    var id: String {
+        switch self {
+        case .profile: return "profile"
+        case .mealForm(let id): return "meal-\(id ?? "new")"
+        case .symptomForm(let id): return "symptom-\(id ?? "new")"
+        case .logEntry: return "log-entry"
+        }
+    }
+}
