+++ GutCheck/GutCheck/Navigation/TabBarAction.swift
@@ -1,0 +1,14 @@
+//
+//  TabBarAction.swift
+//  GutCheck
+//
+//  Actions triggered by the custom tab bar.
+//
+
+// MARK: - Tab Bar Actions
+
+enum TabBarAction {
+    case logMeal
+    case logSymptom
+    case tabTapped(Tab)
+}
